feat(i18n): wrap layout with NextIntlClientProvider + add locale JSONs (en-US, zh-CN)
- Convert layout.tsx to server component that loads messages via next-intl
- Extract client logic into ClientRootLayout.tsx (to keep 'use client' boundary)
- Replace dynamic import in request.ts with static imports + Locale-keyed map
- Add 12 locale JSON files (6 namespaces x 2 locales) for Wave 1 baseline
