+++ frontend/locales/en-US/common.json
@@ -1,0 +1,28 @@
+{
+  "app": {
+    "name": "Meetily",
+    "tagline": "Privacy-first AI meeting assistant"
+  },
+  "nav": {
+    "home": "Home",
+    "meetings": "Meetings",
+    "settings": "Settings",
+    "notes": "Notes"
+  },
+  "actions": {
+    "save": "Save",
+    "cancel": "Cancel",
+    "delete": "Delete",
+    "edit": "Edit",
+    "close": "Close",
+    "confirm": "Confirm",
+    "back": "Back",
+    "next": "Next"
+  },
+  "status": {
+    "loading": "Loading...",
+    "saving": "Saving...",
+    "success": "Success",
+    "failed": "Failed"
+  }
+}
