+++ frontend/src/app/layout.tsx
@@ -1,31 +1,12 @@
-'use client'
-
 import './globals.css'
 import { Source_Sans_3 } from 'next/font/google'
-import Sidebar from '@/components/Sidebar'
-import { SidebarProvider } from '@/components/Sidebar/SidebarProvider'
-import MainContent from '@/components/MainContent'
-import AnalyticsProvider from '@/components/AnalyticsProvider'
-import { Toaster, toast } from 'sonner'
-import "sonner/dist/styles.css"
-import { useState, useEffect, useCallback } from 'react'
-import { listen, UnlistenFn } from '@tauri-apps/api/event'
-import { invoke } from '@tauri-apps/api/core'
-import { TooltipProvider } from '@/components/ui/tooltip'
-import { RecordingStateProvider } from '@/contexts/RecordingStateContext'
-import { OllamaDownloadProvider } from '@/contexts/OllamaDownloadContext'
-import { TranscriptProvider } from '@/contexts/TranscriptContext'
-import { ConfigProvider, useConfig } from '@/contexts/ConfigContext'
-import { OnboardingProvider } from '@/contexts/OnboardingContext'
-import { OnboardingFlow } from '@/components/onboarding'
-import { loadBetaFeatures } from '@/types/betaFeatures'
-import { DownloadProgressToastProvider } from '@/components/shared/DownloadProgressToast'
-import { UpdateCheckProvider } from '@/components/UpdateCheckProvider'
-import { RecordingPostProcessingProvider } from '@/contexts/RecordingPostProcessingProvider'
-import { ImportAudioDialog, ImportDropOverlay } from '@/components/ImportAudio'
-import { ImportDialogProvider } from '@/contexts/ImportDialogContext'
-import { isAudioExtension, getAudioFormatsDisplayList } from '@/constants/audioFormats'
-
+import { NextIntlClientProvider } from 'next-intl'
+import { getLocale, getMessages } from 'next-intl/server'
+import { getStoredLocale } from '@/i18n/request'
+import { LocaleProvider } from '@/hooks/useLocale'
+import ClientRootLayout from './ClientRootLayout'
+import { Toaster } from 'sonner'
+import { DEFAULT_LOCALE, isSupportedLocale, type Locale } from '@/i18n/config'
 
 const sourceSans3 = Source_Sans_3({
   subsets: ['latin'],
@@ -33,249 +14,25 @@ const sourceSans3 = Source_Sans_3({
   variable: '--font-source-sans-3',
 })
 
-// Module-level component — stable reference across RootLayout re-renders.
-// Defined here (not inside RootLayout) so React never sees a new function type
-// on re-render, which would cause unmount/remount and break initialization logic.
-function ConditionalImportDialog({
-  showImportDialog,
-  handleImportDialogClose,
-  importFilePath,
-}: {
-  showImportDialog: boolean;
-  handleImportDialogClose: (open: boolean) => void;
-  importFilePath: string | null;
-}) {
-  const { betaFeatures } = useConfig();
+export { metadata } from './metadata'
 
-  // Only mount ImportAudioDialog (and its hooks/listeners) when feature is enabled
-  if (!betaFeatures.importAndRetranscribe) {
-    return null;
-  }
+export default async function RootLayout({ children }: { children: React.ReactNode }) {
+  // next-intl server helpers + Tauri stored-locale fallback.
+  // PR-12 will replace the getStoredLocale() placeholder with the real
+  // `get_ui_language` Tauri command once the Rust side lands.
+  const intlLocale = await getLocale()
+  const stored = await getStoredLocale()
+  const locale: Locale = isSupportedLocale(stored) ? stored : (isSupportedLocale(intlLocale) ? intlLocale : DEFAULT_LOCALE)
+  const messages = await getMessages()
 
   return (
-    <ImportAudioDialog
-      open={showImportDialog}
-      onOpenChange={handleImportDialogClose}
-      preselectedFile={importFilePath}
-    />
-  );
-}
-
-// export { metadata } from './metadata'
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  const [showOnboarding, setShowOnboarding] = useState(false)
-  const [onboardingCompleted, setOnboardingCompleted] = useState(false)
-
-  // Import audio state
-  const [showDropOverlay, setShowDropOverlay] = useState(false)
-  const [showImportDialog, setShowImportDialog] = useState(false)
-  const [importFilePath, setImportFilePath] = useState<string | null>(null)
-
-  useEffect(() => {
-    // Check onboarding status first
-    invoke<{ completed: boolean } | null>('get_onboarding_status')
-      .then((status) => {
-        const isComplete = status?.completed ?? false
-        setOnboardingCompleted(isComplete)
-
-        if (!isComplete) {
-          console.log('[Layout] Onboarding not completed, showing onboarding flow')
-          setShowOnboarding(true)
-        } else {
-          console.log('[Layout] Onboarding completed, showing main app')
-        }
-      })
-      .catch((error) => {
-        console.error('[Layout] Failed to check onboarding status:', error)
-        // Default to showing onboarding if we can't check
-        setShowOnboarding(true)
-        setOnboardingCompleted(false)
-      })
-  }, [])
-
-  // Disable context menu in production
-  useEffect(() => {
-    if (process.env.NODE_ENV === 'production') {
-      const handleContextMenu = (e: MouseEvent) => e.preventDefault();
-      document.addEventListener('contextmenu', handleContextMenu);
-      return () => document.removeEventListener('contextmenu', handleContextMenu);
-    }
-  }, []);
-  useEffect(() => {
-    // Listen for tray recording toggle request
-    const unlisten = listen('request-recording-toggle', () => {
-      console.log('[Layout] Received request-recording-toggle from tray');
-
-      if (showOnboarding) {
-        toast.error("Please complete setup first", {
-          description: "You need to finish onboarding before you can start recording."
-        });
-      } else {
-        // If in main app, forward to useRecordingStart via window event
-        console.log('[Layout] Forwarding to start-recording-from-sidebar');
-        window.dispatchEvent(new CustomEvent('start-recording-from-sidebar'));
-      }
-    });
-
-    return () => {
-      unlisten.then(fn => fn());
-    };
-  }, [showOnboarding]);
-
-  // Handle file drop for audio import
-  const handleFileDrop = useCallback((paths: string[]) => {
-    // Check if beta features are enabled (read from localStorage directly since we're outside ConfigProvider)
-    const betaFeatures = loadBetaFeatures();
-
-    if (!betaFeatures.importAndRetranscribe) {
-      toast.error('Beta feature disabled', {
-        description: 'Enable "Import Audio & Retranscribe" in Settings > Beta to use this feature.'
-      });
-      return;
-    }
-
-    // Find the first audio file
-    const audioFile = paths.find(p => {
-      const ext = p.split('.').pop()?.toLowerCase();
-      return !!ext && isAudioExtension(ext);
-    });
-
-    if (audioFile) {
-      console.log('[Layout] Audio file dropped:', audioFile);
-      setImportFilePath(audioFile);
-      setShowImportDialog(true);
-    } else if (paths.length > 0) {
-      toast.error('Please drop an audio file', {
-        description: `Supported formats: ${getAudioFormatsDisplayList()}`
-      });
-    }
-  }, []);
-
-  // Listen for drag-drop events
-  useEffect(() => {
-    if (showOnboarding) return; // Don't handle drops during onboarding
-
-    const unlisteners: UnlistenFn[] = [];
-    const cleanedUpRef = { current: false };
-
-    const setupListeners = async () => {
-      // Drag enter/over - show overlay only if beta feature is enabled
-      const unlistenDragEnter = await listen('tauri://drag-enter', () => {
-        if (loadBetaFeatures().importAndRetranscribe) {
-          setShowDropOverlay(true);
-        }
-      });
-      if (cleanedUpRef.current) {
-        unlistenDragEnter();
-        return;
-      }
-      unlisteners.push(unlistenDragEnter);
-
-      // Drag leave - hide overlay
-      const unlistenDragLeave = await listen('tauri://drag-leave', () => {
-        setShowDropOverlay(false);
-      });
-      if (cleanedUpRef.current) {
-        unlistenDragLeave();
-        unlisteners.forEach(u => u());
-        return;
-      }
-      unlisteners.push(unlistenDragLeave);
-
-      // Drop - process files
-      const unlistenDrop = await listen<{ paths: string[] }>('tauri://drag-drop', (event) => {
-        setShowDropOverlay(false);
-        handleFileDrop(event.payload.paths);
-      });
-      if (cleanedUpRef.current) {
-        unlistenDrop();
-        unlisteners.forEach(u => u());
-        return;
-      }
-      unlisteners.push(unlistenDrop);
-    };
-
-    setupListeners();
-
-    return () => {
-      cleanedUpRef.current = true;
-      unlisteners.forEach((unlisten) => unlisten());
-    };
-  }, [showOnboarding, handleFileDrop]);
-
-  // Handle import dialog close
-  const handleImportDialogClose = useCallback((open: boolean) => {
-    setShowImportDialog(open);
-    if (!open) {
-      setImportFilePath(null);
-    }
-  }, []);
-
-  // Handler for ImportDialogProvider - opens import dialog from any child component
-  const handleOpenImportDialog = useCallback((filePath?: string | null) => {
-    setImportFilePath(filePath ?? null);
-    setShowImportDialog(true);
-  }, []);
-
-  const handleOnboardingComplete = () => {
-    console.log('[Layout] Onboarding completed, reloading app')
-    setShowOnboarding(false)
-    setOnboardingCompleted(true)
-    // Optionally reload the window to ensure all state is fresh
-    window.location.reload()
-  }
-
-  return (
-    <html lang="en">
+    <html lang={locale}>
       <body className={`${sourceSans3.variable} font-sans antialiased`}>
-        <AnalyticsProvider>
-          <RecordingStateProvider>
-            <TranscriptProvider>
-              <ConfigProvider>
-                <OllamaDownloadProvider>
-                  <OnboardingProvider>
-                    <UpdateCheckProvider>
-                      <SidebarProvider>
-                        <TooltipProvider>
-                          <RecordingPostProcessingProvider>
-                            <ImportDialogProvider onOpen={handleOpenImportDialog}>
-                              {/* Download progress toast provider - listens for background downloads */}
-                              <DownloadProgressToastProvider />
-
-                              {/* Show onboarding or main app */}
-                              {showOnboarding ? (
-                                <OnboardingFlow onComplete={handleOnboardingComplete} />
-                              ) : (
-                                <div className="flex">
-                                  <Sidebar />
-                                  <MainContent>{children}</MainContent>
-                                </div>
-                              )}
-                              {/* Import audio overlay and dialog */}
-                              <ImportDropOverlay visible={showDropOverlay} />
-                              <ConditionalImportDialog
-                                showImportDialog={showImportDialog}
-                                handleImportDialogClose={handleImportDialogClose}
-                                importFilePath={importFilePath}
-                              />
-                            </ImportDialogProvider>
-                          </RecordingPostProcessingProvider>
-                        </TooltipProvider>
-                      </SidebarProvider>
-                    </UpdateCheckProvider>
-                  </OnboardingProvider>
-
-                </OllamaDownloadProvider>
-              </ConfigProvider>
-            </TranscriptProvider>
-          </RecordingStateProvider>
-        </AnalyticsProvider>
-
+        <NextIntlClientProvider locale={locale} messages={messages}>
+          <LocaleProvider initial={locale}>
+            <ClientRootLayout>{children}</ClientRootLayout>
+          </LocaleProvider>
+        </NextIntlClientProvider>
         <Toaster position="bottom-center" richColors closeButton />
       </body>
     </html>
